refactor(hipsr): move ConvertToLLVM interface into the dialect IR library
hipsr.constant's lowering pattern already lives in the IR library, so move the
ConvertToLLVMPatternInterface implementation and its registration there too
(HipsrDialect.cpp) and delete the now-empty lib/Conversion/HipsrToLLVM library.
The registration declaration moves to HipsrDialect.h; callers include it
already. hip, convert-hip-to-llvm, and test semantics are unchanged.
